+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - server: provide option to explicitly recreate event triggers for sources in the `reload_metadata` API
 - server: fix `gen_hasura_uuid` migration to be idempotent, so that it doesn't fail if the database is already initialised with source migrations.
 - server: fix mssql `table_by_pk` query returning empty array (fix #7784)
+- server: fix BigQuery queries failing with more than one array relationship
 - console: add comments to tracked functions
 - console: add select all columns option while selecting the columns in event triggers
 - console: add request transforms for events